+++ CHANGELOG.md
@@ -5,8 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.0.0-beta.12] - 2018-01-27
 ### Added
 - Added Patreon link in Contribute admin page
-- Added Theme Code Injection feature
+- Added Theme Code Injection functionality
 - Added Theme CSS Injection code minification
+- Added Page Delete functionality
 
 ### Fixed
 - Fixed root admin refresh token fail
@@ -14,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Moved Insert Media button in Markdown editor
+- Use semver for DB migrations ordering
 
 ## [2.0.0-beta.11] - 2018-01-20
 - First beta release