+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fixed root admin refresh token fail
 - Fixed error page metadata title warning
+- Fixed telemetry
 
 ### Changed
 - Moved Insert Media button in Markdown editor