+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed root admin refresh token fail
 - Fixed error page metadata title warning
 - Fixed telemetry
+- Await page render job to complete before resolving
 
 ### Changed
 - Moved Insert Media button in Markdown editor