+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [2.0.0-beta.12] - 2018-01-27
+## [2.0.0-beta.XX] - 2018-XX-XX
 ### Added
 - Added Patreon link in Contribute admin page
 - Added Theme Code Injection functionality
 - Added Theme CSS Injection code minification
 - Added Page Delete functionality
+- Dev locale .yml files in `server/locales` are now loaded
 
 ### Fixed
 - Fixed root admin refresh token fail